♻️ Moves <VisualRoot /> into <App />
